+++ include/sadfs/msgs/master_messages.hpp
@@ -1,0 +1,65 @@
+#ifndef SADFS_MSGS_MASTER_MESSAGES_HPP
+#define SADFS_MSGS_MASTER_MESSAGES_HPP
+
+// sadfs-specific includes
+#include <sadfs/msgs/common.hpp>
+#include <sadfs/proto/master.pb.h>
+
+// standard includes
+#include <cstddef> // std::size_t
+#include <string>
+
+namespace sadfs { namespace msgs { namespace master {
+using control_message = proto::master::control_message;
+
+// enumerates types of raw messages
+enum class msg_type
+{
+	unknown,
+	chunk_location_request,
+};
+
+class chunk_location_request
+{
+public:
+	chunk_location_request() = default;
+	chunk_location_request(msgs::io_type, std::string const& filename,
+	                       std::size_t chunk_number);
+
+	msgs::io_type      io_type()     const;
+	std::string const& filename()    const;
+	std::size_t        chunk_number() const;
+	
+	inline static msg_type type{msg_type::chunk_location_request};
+private:
+	proto::master::chunk_location_request protobuf_{};
+
+	// provide embed/extract functions access to private members
+	friend bool embed(chunk_location_request const&, control_message&);
+	friend bool extract(chunk_location_request&, control_message const&);
+};
+
+// declarations
+bool embed(chunk_location_request const&, control_message&);
+bool extract(chunk_location_request&, control_message const&);
+
+// ==================================================================
+//                     inline function definitions
+// ==================================================================
+inline std::string const& chunk_location_request::
+filename() const
+{
+	return protobuf_.filename();
+}
+
+inline std::size_t chunk_location_request::
+chunk_number() const
+{
+	return protobuf_.chunk_number();
+}
+
+} // master namespace
+} // msgs namespace
+} // sadfs namespace
+
+#endif // SADFS_MSGS_MASTER_MESSAGES_HPP
