--- src/msgs/messages.cpp
@@ -1,170 +1,0 @@
-/* implementation of control messages */
-// sadfs-specific includes
-#include <sadfs/msgs/messages.hpp>
-
-// standard includes
-#include <unordered_map>
-
-namespace sadfs { namespace msgs {
-
-using io_type_map = std::unordered_map<proto::io_type,
-                                            io_type>;
-using proto_io_type_map = std::unordered_map<io_type,
-                                            proto::io_type>;
-namespace {
-
-auto const proto_io_type_lookup = proto_io_type_map
-{
-	{io_type::unknown, proto::io_type::UNKNOWN},
-	{io_type::read, proto::io_type::READ},
-	{io_type::write, proto::io_type::WRITE},
-};
-
-auto const io_type_lookup = io_type_map
-{
-	{proto::io_type::UNKNOWN, io_type::unknown},
-	{proto::io_type::READ, io_type::read},
-	{proto::io_type::WRITE, io_type::write},
-};
-
-} // unnamed namespace
-
-// ========================================================
-//                      master namespace
-// ========================================================
-namespace master {
-
-using proto::master::control_message;
-using MsgCase = control_message::MsgCase;
-using msg_type_map = std::unordered_map<MsgCase, msg_type>;
-namespace {
-
-auto const msg_type_lookup = msg_type_map
-{
-	{MsgCase::MSG_NOT_SET,       msg_type::unknown},
-	{MsgCase::kChunkLocationReq, msg_type::chunk_location_request},
-};
-
-} // unnamed namespace
-
-/* ========================================================
- *                       chunk_location_request
- * ========================================================
- */
-chunk_location_request::
-chunk_location_request(msgs::io_type type,
-                       std::string const& filename,
-                       std::size_t chunk_number)
-{
-	protobuf_.set_type(proto_io_type_lookup.at(type));
-	protobuf_.set_filename(filename);
-	protobuf_.set_chunk_number(chunk_number);
-}
-
-io_type chunk_location_request::
-io_type() const
-{
-	return io_type_lookup.at(protobuf_.type());
-}
-
-// embeds a raw message into control message that is
-// (typically) sent over the wire
-bool
-embed(chunk_location_request const& req, control_message& cm)
-{
-	// should this be in a try-catch block?
-	// msg.mutable_chunk_location_req() can throw if heap allocation fails
-	*cm.mutable_chunk_location_req() = req.protobuf_;
-	return true;
-}
-
-// extracts a raw message from the control message that is
-// (typically) received over the wire
-bool
-extract(chunk_location_request& req, control_message const& cm)
-{
-	if (msg_type_lookup.at(cm.msg_case()) != chunk_location_request::type)
-	{
-		// cannot extract a msg that doesn't exist
-		return false;
-	}
-
-	// read chunk_location_request from control_message
-	req.protobuf_ = cm.chunk_location_req();
-	return true;
-}
-
-} // master namespace
-
-// ========================================================
-//                      chunk namespace
-// ========================================================
-namespace chunk {
-
-using proto::chunk::control_message;
-using MsgCase = control_message::MsgCase;
-using msg_type_map = std::unordered_map<MsgCase, msg_type>;
-namespace {
-
-auto const msg_type_lookup = msg_type_map
-{
-	{MsgCase::MSG_NOT_SET, msg_type::unknown},
-	{MsgCase::kChunkReq,   msg_type::chunk_request},
-};
-
-} // unnamed namespace
-
-/* ========================================================
- *                       chunk_request
- * ========================================================
- */
-chunk_request::
-chunk_request(msgs::io_type type, std::size_t chunk_id)
-{
-	protobuf_.set_type(proto_io_type_lookup.at(type));
-	protobuf_.set_chunk_id(chunk_id);
-}
-
-std::size_t chunk_request::
-chunk_id() const
-{
-	return protobuf_.chunk_id();
-}
-
-io_type chunk_request::
-io_type() const
-{
-	return io_type_lookup.at(protobuf_.type());
-}
-
-// embeds a raw message into a control message that is
-// (typically) sent over the wire
-bool
-embed(chunk_request const& req, control_message& cm)
-{
-	// should this be in a try-catch block?
-	// msg.mutable_chunk_req() can throw if heap allocation fails
-	*cm.mutable_chunk_req() = req.protobuf_;
-	return true;
-}
-
-// extracts a raw message from the control message that is
-// (typically) received over the wire
-bool
-extract(chunk_request& req, control_message const& cm)
-{
-	if (msg_type_lookup.at(cm.msg_case()) != chunk_request::type)
-	{
-		// cannot extract a msg that doesn't exist
-		return false;
-	}
-
-	// read chunk_request from control_message
-	req.protobuf_ = cm.chunk_req();
-	return true;
-}
-
-} // chunk namespace
-
-} // namespace msgs
-} // namespace sadfs
